rename hpp files. add main file to docs.
