+++ TronGame/CMakeLists.txt
@@ -1,7 +1,3 @@
-# cmake_minimum_required(VERSION 3.5)
-
-# project(TronGame)
-
 file(COPY tronData DESTINATION ${CMAKE_BINARY_DIR})
 file(COPY tronData DESTINATION ${CMAKE_BINARY_DIR}/TronGame)
 
@@ -44,23 +40,3 @@ target_link_libraries(
         sfml-system
         sfml-network
 )
-
-# add_executable(
-#         ctest
-#         tests.cpp
-# )
-
-# target_link_libraries(ctest 
-#         PUBLIC
-#         doctest
-#         Tron_lib
-#         sfml-graphics
-#         sfml-window
-#         sfml-system
-#         sfml-network
-# )
-
-# cmake_minimum_required(VERSION 3.5)
-
-
-# project(TicTacToe)
