+++ main.cpp
@@ -1,10 +1,21 @@
+/*! \file main.cpp
+\brief Main file for project.
+    GAll the games are called from here.
+*/
+
 #include <iostream>
-#include <string>
 
 #include "TicTacToe/server.h"
 #include "TicTacToe/client.h"
 #include "TronGame/game.h"
 
+/**
+* @brief Entry point
+*
+* Execution of the project starts here.
+*
+* @return Program exit status
+*/
 int main() {
     using std::cin;
     using std::cout;
@@ -14,7 +25,7 @@ int main() {
     cin >> game;
     if (game == 1) {
         char typeOfConnection;
-        cout << "Enter type of connection. ";
+        cout << "Enter type of connection. (s - server, c - client)";
         cin >> typeOfConnection;
         if(typeOfConnection == 's') {
             // server's logic
@@ -23,8 +34,8 @@ int main() {
             // client's logic
             client(55002, "127.0.0.1", 55001);
         }
-    } else {
+    } else if (game == 2) {
         gameLoop();
     }
-
+    return 0;
 }
